Make default connect timeout longer
The default 5 seconds is too short for wifi connects to return if
there was an error trying to connect or not. This resulted in a
timeout, followed by an invalid_key when the creds for the ssid were
wrong.

Increasing the timeout (as the connman documentation suggests) solves
this problem